修复了level2无法正常结束的问题

+++ PrismGC/Keil/src/BlockMap.h
@@ -19,11 +19,11 @@
 
 typedef struct
 {
-    int32_t i;
-    int32_t j;
+    int8_t i;
+    int8_t j;
 } MapCoord;
 
-MapCoord _MapCoord(int32_t i,int32_t j);
+MapCoord _MapCoord(int8_t i,int8_t j);
 MapCoord MapCoordPlus(MapCoord a,MapCoord b);
 MapCoord MapCoordMinus(MapCoord a,MapCoord b);
 uint8_t MapCoordEqual(MapCoord a,MapCoord b);
